Screen:
 * Screen('GetWindowInfo') now returns if window is a fullscreen
   window: For this, a new 'specialflags' flag kPsychIsFullscreenWindow
   gets set and stored on all OS backends if an onscreen window is
   a fullscreen window. The function also exposes the 'specialflags'
   for future use.

 * Screen('Computer') on Windows now returns 'version' field with
   OS version major.minor.build and an optional "ServicePack" string.

 * New conserveVRAM flags to enforce or prevent use of Aero DWM on
   Vista et al., and to enforce or prevent use of SetForeGroundWindow()
   and SetFocus() calls.

 * Documentation updates in internal code and updates of user visible
   helptexts.

 * Reenable support for 16bpc and 32bpc float framebuffers for video
   scanout. Windows-7 supports 16bpc float scanout on some 
   DirectX10/11 hardware, OpenGL may expose this as well in the future.

 * Compile-fixes on Windows.

Screen for Windows Vista/Windows-7:

 * Automatically decide if the DWM should be enabled, disabled or
   left in its current state ("don't care"), depending on OS,
   presence of fullscreen/non-fullscreen windows, PsychDebugWindow...
   settings for transparency. Also decide if SetFocus(),SetForeground()
   should be called or not on a new onscreen window. Automatic can be
   overriden with conserveVRAM flags mentioned above:

   1. Fullscreen windows without transparency ("regular fullscreen windows")
      will disable the DWM and SetFocus() et al. to provide optimal stimulus
      onset timing/performance and accurate beamposition timestamping with
      hardware VSync'ed, pageflipped doublebuffer-swaps.

   2. Transparent ("debug mode") fullscreen windows will enable the DWM and
      prevent SetFocus() et al. --> Only for debugging convenience, not
      useful for precise timing.

   3. Windowed windows don't SetFocus and don't care about DWM state.

   4. Manual enforce/prevent overrides exist for any of these settings.

 -> All attempts at getting good timing or timestamping with DWM enabled
    were total failures. Needs to be off for reliable timing and performance.

 -> At least we can get full redraw rate (redraw fps = videorefresh Hz) with
    all the new PsychSetPresentParameters() code disabled, even if the onset
    timing is wrong and grossly delayed due to composition lag. Not good enough
    for running experiment sessions, but good enough for code development,
    debugging and initial testing, and a nice convenience feature.

 -> Need to SetFocus() and SetForeGround(), otherwise the DWM may auto-enable
    and interfere, or the GPU will resort to doublebufferswaps that are triggered
    in the VSync interrupt ISR routine, with potentially high delay, and are
    executed via framebuffer BitBlt's instead of pageflipping.

    This can cause severe tearing artifacts and problems. The need for SetFocus()
    is a hard requirement of the implementation of Vista et al. Direct X
    graphics kernel subsystem, so this is unavoidable to my knowledge :-(

Soo, stimulus onset timing and timestamping of high quality can be done on
Vista et al. this way (tests on dualdisplay setups pending). Problem here is
that SetFocus() effectively disables our CharAvail()/GetChar() functions, because
it steals keyboard input focus from Matlab and Octave.

-> Need a new implementation to solve this. Prototype code is implemented in
PsychWindowGlue to find the hwnd of the Matlab/Octave main window by enumeration
and pattern-matching. Works reasonably well to get a window handle hwnd.

-> We try to resubmit (PostEvent()) any character keystroke event from within
our own onscreen windows event queue handler to the event queue of the Matlab/
Octave windows, thereby injecting the received keystrokes into the runtime,
in order to restore GetChar/CharAvail functionality.

-> The current code doesn't work properly yet. On Octave it reliably submits
most keystrokes, but repeated characters vanish (E.g., "Otter" becomes "Oter"
as the repeated 2nd 't' character disappears). On Matlab, simple prototyping
code works, but the actual (almost identical) code in our routines completely
fails to inject keystrokes, although they are received by Matlabs event queue.
So far completely unclear why it doesn't work.



git-svn-id: https://psychtoolbox-3.googlecode.com/svn/trunk@1541 f5efbe48-b203-0410-9429-fc5e098e5a54

+++ PsychSourceGL/Source/Common/Screen/PsychVideoCaptureSupportLibARVideo.c
@@ -586,9 +586,9 @@ int PsychARVideoCaptureRate(int capturehandle, double capturerate, int dropframe
 				}
 				
 				if (capdev->nrframes>0) capdev->avg_decompresstime/= (double) capdev->nrframes;
-				printf("PTB-INFO: Average time spent in video decompressor (waiting/polling for new frames) was %lf milliseconds.\n", capdev->avg_decompresstime * 1000.0f);
+				printf("PTB-INFO: Average time spent in video decompressor (waiting/polling for new frames) was %f milliseconds.\n", (float) capdev->avg_decompresstime * 1000.0f);
 				if (capdev->nrgfxframes>0) capdev->avg_gfxtime/= (double) capdev->nrgfxframes;
-				printf("PTB-INFO: Average time spent in GetCapturedImage (intensity calculation Video->OpenGL texture conversion) was %lf milliseconds.\n", capdev->avg_gfxtime * 1000.0f);
+				printf("PTB-INFO: Average time spent in GetCapturedImage (intensity calculation Video->OpenGL texture conversion) was %f milliseconds.\n",  (float) capdev->avg_gfxtime * 1000.0f);
 			}
 		}
 	}
